app/requirements.txt
google-cloud-datastore==1.7.0
flask==1.0.2